24_07_29 벡엔드 배포버전 업데이트

+++ src/main/java/io/paioneer/nain/interview/jpa/repository/InterviewRepository.java
@@ -15,10 +15,10 @@ public interface InterviewRepository extends JpaRepository<InterviewEntity, Long
     Page<InterviewEntity> findAllByMemberNo(Long memberNo, Pageable pageable);
 
     @Query("SELECT COUNT(s) FROM InterviewEntity s WHERE s.videoScore < :score AND s.voiceScore < :score")
-    int countScoresLowerThan(@Param("score") int score);
+    int countScoresLowerThan(@Param("score") double score);
 
     @Query("SELECT COUNT(s) FROM InterviewEntity s WHERE s.videoScore > :score AND s.voiceScore > :score")
-    int countScoresHigherThan(@Param("score") int score);
+    int countScoresHigherThan(@Param("score") double score);
 
     @Query("SELECT COUNT(s) FROM InterviewEntity s")
     int countTotalScores();
